PARQUET-1078: Add option to coerce Arrow timestamps to a particular unit
See ARROW-622 and ARROW-1076. We have been coercing timestamp units on ingest to Arrow format, rather than letting the data be whatever it is and converting only when writing to Parquet. This adds an option to coerce via the `ArrowWriterProperties`, and we also will return error Status if casting to a lower-resolution unit would lose data. We can later add an option to allow unsafe casts should that be desired.

Author: Wes McKinney <wes.mckinney@twosigma.com>

Closes #380 from wesm/PARQUET-1078 and squashes the following commits:

2bb6d9f [Wes McKinney] Code reviews. Use CapWords for lambda functions
0d5ff4d [Wes McKinney] Account for offset in WriteTimestampsCoerce
17402cb [Wes McKinney] Add invalid value in null slot
94ecc65 [Wes McKinney] Only check for data loss when values are not null
7069757 [Wes McKinney] Implement timestamp unit coercion, raise when losing precision
b97b78b [Wes McKinney] Initial tests for timestamp coercion
f5ddb68 [Wes McKinney] Add schema logic for timestamp coercion
69bb26a [Wes McKinney] Refactor schema functions to enable more Arrow writer properties. Add timestamp coercion functions

Change-Id: I396a56a349e736a70d5c1ba030c20af2b7ef4cd0