ARROW-2434: [Rust] Add windows support
I'm not that familiar with JIRA but I believe that I do not have the permission to self-assign this issue.  @andygrove maybe you can assign to me on JIRA please (or let me know if I should be able to self-assign).

I had to expose `_aligned_malloc` and `_aligned_free` but I'm thinking that this should be in _libc_.  This is just to get windows support building.  I'll submit a PR to libc and open another JIRA to clean things up.

Author: Paddy <paddyhoran@hotmail.com>

Closes #1873 from paddyhoran/master and squashes the following commits:

7385d31 <Paddy> Removing the use of `if cfg!` macro to fix non windows builds.
35fa405 <Paddy> Fixed formatting issues.
85ea7d9 <Paddy> Fixes ARROW-2434