ARROW-2545: [Python] Link against required system libraries
Python can require system libraries such as "libutil".  When linking dynamically against libpythonXX.so, this is not a problem since those libraries are automatically depended on.  When linking statically
against libpythonXX.a (this is the case when Python itself is statically linked and therefore does not provide a .so), you need to specify those libraries explicitly.

Author: Antoine Pitrou <antoine@python.org>

Closes #2005 from pitrou/ARROW-2545-python-other-libs and squashes the following commits:

773ad168 <Antoine Pitrou> ARROW-2545:  Link against required system libraries