ARROW-2594: [Java] When realloc Vectors, zero out all unfilled bytes of new buffer
Currently when reallocating vectors, only the second half of the new buffer will be zeroed out assuming that it is doubled from the previous buffer and the first half is already populated or cleaned.  This isn't the case if the vector had been cleared and the buffer is empty causing incorrect values in the new buffer if it was recycled from an old one.

Added a new test with a ListVector that should reuse a previous buffer after being cleared.

Author: Bryan Cutler <cutlerb@gmail.com>

Closes #2054 from BryanCutler/java-vector-realloc-clear-buffer-ARROW-2594 and squashes the following commits:

28b8095 <Bryan Cutler> added a comment about clear
be3ee8f <Bryan Cutler> remove extra spaces
5a39790 <Bryan Cutler> zero out any newly allocated buffer bytes