ARROW-3357: [Rust] Add a mutable buffer implementation
Currently we only have a immutable buffer implementation (`Buffer`). To implement various array builders, it's better to also have a mutable version of this so that we can keep appending data to a buffer. This PR adds a `MutableBuffer` struct which is growable and can be mutated. A mutable buffer can also be consumed and converted into a immutable buffer.

Author: Chao Sun <sunchao@uber.com>
Author: Krisztián Szűcs <szucs.krisztian@gmail.com>

Closes #2658 from sunchao/ARROW-3357 and squashes the following commits:

3aaa8252 <Krisztián Szűcs> rustfmt
a182400c <Chao Sun> Fix lint
7893ff1c <Chao Sun> ARROW-3357:  Add a mutable buffer implementation