ARROW-4070: [C++] Enable use of ARROW_BOOST_VENDORED with ninja-build
It seems that ninja-build is a lot stricter about the dependency graph -- it seeks the root dependency of the `boost_*_static` libraries and finds targets (the absolute paths to the static libraries) that it doesn't know how to build. Setting these as the BUILD_BYPRODUCTS of the ExternalProject fixes the issue.

I need this fix in ARROW-3803 so I'm going to cherry pick it there, and I can rebase later

Author: Wes McKinney <wesm+git@apache.org>

Closes #3217 from wesm/ARROW-4070 and squashes the following commits:

aac135daa <Wes McKinney> Use static library paths as BOOST_BUILD_PRODUCTS so that ninja-build can understand the dependency graph