ARROW-2592: [Python] Add "ignore_metadata" option to Table.to_pandas
This option circumvents the index reconstruction logic if there is `'pandas'` metadata. This can also be achieved using `table.cast(table.schema.remove_metadata()).to_pandas()`, but this makes it more obvious and discoverable to users.

A user had an issue reading a Parquet file with some old metadata that we are no longer able to correctly process.

Author: Wes McKinney <wesm+git@apache.org>

Closes #3239 from wesm/ARROW-2592 and squashes the following commits:

6abe6ac0 <Wes McKinney> Unit test for ignore_metadata option
35476952 <Wes McKinney> Test stub
701697b7 <Wes McKinney> Add ignore_metadata option to Table.to_pandas