ARROW-5429: [Java] Provide alternative buffer allocation policy
The current buffer allocation policy works like this:

1. If the requested buffer size is greater than or equal to the chunk size, the buffer size will be as is.
2. If the requested size is within the chunk size, the buffer size will be rounded to the next power of 2.

This policy can lead to waste of memory in some cases. For example, if we request a buffer of size 10MB, Arrow will round the buffer size to 16 MB. If we only need 10 MB, this will lead to a waste of (16 - 10) / 10 = 60% of memory.

So in this proposal, we provide another policy: the rounded buffer size must be a multiple of some memory unit, like (32 KB). This policy has two benefits:

1. The wasted memory cannot exceed one memory unit (32 KB), which is much smaller than the power-of-two policy.
2. This is the memory allocation policy adopted by some computation engines (e.g. Apache Flink).

Author: liyafan82 <fan_li_ya@foxmail.com>

Closes #4400 from liyafan82/fly_5429 and squashes the following commits:

6b44dce2 <liyafan82> Merge branch 'master' into fly_5429
ed795105 <liyafan82> Resolve comments and evaluate performance
578d5fde <liyafan82> Resolve comments
a6840e15 <liyafan82> Remove useless import
36bbeb14 <liyafan82> Provide alternative buffer allocation policy