ARROW-5495: [C++] Update some dependency URLs from http to https
HTTPS is more secure.

I will leave this patch up in case anyone else wants to look for HTTP URLs to change to HTTPS. It's a bit hard to find them because of the `http://` in the Apache license header. Maybe we should change our Apache license headers to use https?

Author: Wes McKinney <wesm+git@apache.org>

Closes #4463 from wesm/use-https and squashes the following commits:

159dddc2a <Wes McKinney> Update some URLs from http to https