ARROW-5579: [Java] shade flatbuffer dependency
Related to [ARROW-5579](https://issues.apache.org/jira/browse/ARROW-5579).

After some discussion with the Flatbuffers maintainer, it appears that FB generated code is not guaranteed to be compatible with any other version of the runtime library other than the exact same version of the flatc used to compile it.

This makes depending on flatbuffers in a library (like arrow) quite risky, as if an app depends on any other version of FB, either directly or transitively, it's likely the versions will clash at some point and you'll see undefined behaviour at runtime.

Shading the dependency looks to me the best way to avoid this.

Author: tianchen <niki.lj@alibaba-inc.com>

Closes #4540 from tianchen92/ARROW-5579 and squashes the following commits:

fbd0c7176 <tianchen> ARROW-5579:  shade flatbuffer dependency