ARROW-5555: [R] Add install_arrow() function to assist the user in obtaining C++ runtime libraries
Note that this function doesn't install anything: it provides context-dependent advice and for some platforms recommends commands to run. In future versions, we may want to make it do more, but this is a useful starting point.

In addition to this function, this patch also:

* Updates the README with installation instructions, including for Windows (assuming https://github.com/apache/arrow/pull/4622)
* Adds a Makefile to facilitate local package dev, building, testing, etc. (https://issues.apache.org/jira/browse/ARROW-5328)
* Fixes an UTF-8 issue in the DESCRIPTION
* A few other cleanups

Author: Neal Richardson <neal.p.richardson@gmail.com>

Closes #4654 from nealrichardson/r-install-arrow and squashes the following commits:

90c2e5e8 <Neal Richardson> Simplify installation instructions
a2145629 <Neal Richardson> Add ARROW_R_DEV to Makefile
b2870489 <Neal Richardson> if (
60150149 <Neal Richardson> Update authors, including ASCII fix
3327cb1b <Neal Richardson> Implement messages for install_arrow, write tests
af34401c <Neal Richardson> Sketch of install_arrow()
4c756bb3 <Neal Richardson> Update installation docs; add Makefile for CLI-oriented folks