ARROW-5427: [Python] pandas conversion preserve_index=True to force RangeIndex serialization
https://issues.apache.org/jira/browse/ARROW-5427

This proposes to let `preserve_index=True` force the index to be a column, also RangeIndexes. The default is unchanged (only now codified as None) to have store a RangeIndex as metadata only. But this gives the possibility to force serialization if consistent results are required (or need to match a specified schema, cfr https://issues.apache.org/jira/browse/ARROW-5220)

Author: Joris Van den Bossche <jorisvandenbossche@gmail.com>

Closes #4651 from jorisvandenbossche/ARROW-5427-preserve-index and squashes the following commits:

56efb38ba <Joris Van den Bossche> ARROW-5427:  pandas conversion preserve_index=True to force RangeIndex serialization