ARROW-5973: [Java] Variable width vectors' get methods should return null when the underlying data is null
For variable-width vectors (VarCharVector and VarBinaryVector), when the validity bit is not set, it means the underlying data is null, so the get method should return null.

However, the current implementation throws an IllegalStateException when NULL_CHECKING_ENABLED is set, or returns an empty array when the flag is clear.

Maybe the purpose of this design is to be consistent with fixed-width vectors. However, the scenario is different: fixed-width vectors (e.g. IntVector) throw an IllegalStateException, simply because the primitive types are non-nullable.

Author: liyafan82 <fan_li_ya@foxmail.com>

Closes #4901 from liyafan82/fly_0717_varget and squashes the following commits:

8fe83f751 <liyafan82>  Variable width vectors' get methods should return null when the underlying data is null