ARROW-5898: [Java] Provide functionality to efficiently compute hash code for arbitrary memory segment
This issue adds a functionality to efficiently compute  the hash code for a consecutive memory region. This functionality is important in practical scenarios because it helps:

*Avoid unnecessary memory copy.
*Avoid repeated conversions between Java objects & Arrow buffers.

Since the algorithm for calculating hash code has  significant performance implications, we need to design an interface so that different algorithms can be easily introduces as plug-ins.

Author: liyafan82 <fan_li_ya@foxmail.com>

Closes #4844 from liyafan82/fly_0710_hash1 and squashes the following commits:

b1b6f78d0 <liyafan82>  Provide functionality to efficiently compute hash code for arbitrary memory segment