ARROW-1184: [Java] Dictionary.equals is not working correctly
Related to [ARROW-1184](https://issues.apache.org/jira/browse/ARROW-1184).
The Dictionary.equals method does not return True when the dictionaries are equal. This is because equals is not implemented for FieldVector and so that comparison defaults to comparing the two objects only and not the vector data.

Closes #4843 from tianchen92/ARROW-1184 and squashes the following commits:

3511857 <tianchen> revert
527a9d7 <tianchen> add TODO
0c69588 <tianchen> fix equals
1f41366 <tianchen> fix build
87b7f66 <tianchen> fix
3fe9389 <tianchen> use Validator logic in Dictionary#equals and add UT
be74915 <tianchen> move UT
314d697 <tianchen> ARROW-1184:  Dictionary.equals is not working correctly

Authored-by: tianchen <niki.lj@alibaba-inc.com>
Signed-off-by: Bryan Cutler <cutlerb@gmail.com>