ARROW-5028: [Python] Avoid malformed ListArray types caused by reaching StringBuilder capacity when converting from Python sequence
This fixes the root cause of ARROW-5028 was was malformed `list<string>` being produced by `pyarrow.array` when the contents of the child builder exceed the 2GB capacity limit of a StringArray. Internally, the child `SeqConverter` was correctly yielding multiple chunks, but this was passing unbeknownst to the parent `ListConverter`. We'll have to do some follow up work to produce chunked ListArray in the future but this at least raises an exception for now so that we aren't creating malformed arrays.

Closes #5108 from wesm/ARROW-5028 and squashes the following commits:

7076e4d5d <Wes McKinney> Check List child builders for chunk overflow when converting from Python sequences

Authored-by: Wes McKinney <wesm+git@apache.org>
Signed-off-by: Wes McKinney <wesm+git@apache.org>