ARROW-5562: [C++][Parquet] Write negative zero or small epsilons as positive zero when computing Parquet statistics
The OP reported a bug where we could have the "max" statistic equal to `-0.0` while the min is `+0.0`, which causes an error in some processing frameworks like Hive.

I also added logic to zero out small epsilons. If this offends anyone, can you suggest an alternative approach to fixing this?

Closes #5375 from wesm/ARROW-5562 and squashes the following commits:

284c436c8 <Wes McKinney> fix typo
11d89904d <Wes McKinney> Write negative zero or small epsilons as positive zero

Authored-by: Wes McKinney <wesm+git@apache.org>
Signed-off-by: Benjamin Kietzman <bengilgit@gmail.com>