ARROW-5086: [Python][Parquet] Opt in to file memory-mapping when reading Parquet files rather than opting out
I was surprised to find that accessing a memory map linearly increases resident RAM usage. Had I know that I might not have made the default for the `memory_map` argument True. It does improve performance by 10-15% as far as I can tell with the downside of increased RSS usage. This seems counterintuitive to users, so I think it is probably better if we disable it by default even though performance will be _slightly_ worse.

Closes #5433 from wesm/ARROW-5086 and squashes the following commits:

0a8358ad0 <Wes McKinney> Opt in to memory mapping when reading Parquet files rather than opting out

Authored-by: Wes McKinney <wesm+git@apache.org>
Signed-off-by: Wes McKinney <wesm+git@apache.org>