ARROW-6722: [Java] Provide a uniform way to get vector name
Currently, the getName method is defined in BaseValueVector, as an abstract class. However, some vector does not extend the BaseValueVector, like StructVector, UnionVector, ZeroVector.

In this issue, we move the method to ValueVector interface, the base interface for all vectors.

This makes it easier to get a vector's name without checking its type.

Closes #5525 from liyafan82/fly_0927_name and squashes the following commits:

8e3c2897a <liyafan82>  Provide a uniform way to get vector name

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Wes McKinney <wesm+git@apache.org>