ARROW-6695: [Rust] [DataFusion] Remove legacy code for executing logical plan
As discussed in the mailing list, this PR removes the legacy code that executed queries directly from the LogicalPlan. This code is not needed now that the new query execution code based on a physical query plan has reached feature parity.

We should merge https://github.com/apache/arrow/pull/5582 first, and rebase.

Closes #5583 from andygrove/ARROW-6695 and squashes the following commits:

6105ea96e <Andy Grove> Rebase
195ed7607 <Andy Grove> Remove legacy relations
8a7a0d276 <Andy Grove> Remove legacy code that executed queries directly from a logical plan

Authored-by: Andy Grove <andygrove73@gmail.com>
Signed-off-by: Andy Grove <andygrove73@gmail.com>