ARROW-7072: [Java] Support concating validity bits efficiently
For scenarios when we need to concate vectors (like the scenario in ARROW-7048, and delta dictionary), we need a way to concat validity bits.

Currently, we have bit level API to read/write individual validity bit. However, it is not efficient , and we need a way to copy more bits at a time.

Closes #5782 from liyafan82/fly_1106_bits and squashes the following commits:

70c11737f <liyafan82>  Resolve comments
d7bcfaaa3 <liyafan82>  Further improve the performance
4f1b9af8b <liyafan82>  Resolve comments
591cf74e4 <liyafan82>  Support concating validity bits efficiently

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>