ARROW-7685: [Developer] Add support for GitHub Actions to Crossbow
If we use GitHub Actions, we don't need to set additional secret
variables to crossbow repository because we can use built-in
GITHUB_TOKEN secret variable.

We'll be able to use self-hosted runner with GitHub Actions. It means
that we'll be able to build .deb/.rpm packages on ARM machine instead
on QEMU.

linux-packages tasks are migrated to GitHub Actions from Azure
Pipelines as an example.

Closes #6286 from kou/crossbow-github-actions and squashes the following commits:

3a75ac42a <Krisztián Szűcs> set workflow name
35dd859f7 <Krisztián Szűcs> use CROSSBOW_GITHUB_TOKEN secret
75c574a5f <Sutou Kouhei>  Add support for GitHub Actions to Crossbow

Lead-authored-by: Sutou Kouhei <kou@clear-code.com>
Co-authored-by: Krisztián Szűcs <szucs.krisztian@gmail.com>
Signed-off-by: Krisztián Szűcs <szucs.krisztian@gmail.com>