ARROW-7975: [C++] Preserve intended buffer size by default when writing to IPC format
I stumbled into this tangentially. While padding to a minimum of an 8-byte multiple is a requirement of the binary protocol, including the padding in the metadata causes buffer sizes to be modified. In some cases, the intention of the writer is for the receiver to receive padded buffers, but including the padding unconditionally leaves no possibility of recovering the buffer size prior to producing the IPC message. I think it would be better to give the writer the option of what to do. I opened ARROW-7976 about adding an option to make this behavior configurable.

Another possibility of course is to implement an option and have "include padding" turned on by default

To be clear in case there is concern, this change has no backward or forward compatibility implications.

Closes #6513 from wesm/ARROW-7975 and squashes the following commits:

df4e15e18 <Wes McKinney> Add unit test to check that buffer sizes pass through IPC unmodified
2d9a97ea1 <Wes McKinney> Do not include padding in Buffer IPC metadata

Authored-by: Wes McKinney <wesm+git@apache.org>
Signed-off-by: Wes McKinney <wesm+git@apache.org>