ARROW-7988: [R] Fix on.exit calls in reticulate bindings
I missed this/didn't closely review after the refactoring that pulled the schema out of the array object. `on.exit()` overwrites the existing finalizer (more recent versions of R have added an `add = TRUE` argument that lets you append, but we still support older versions) so we should only call it once in each function. As the code currently stands, we wouldn't actually be calling `delete_arrow_schema(schema_ptr)`.

Closes #6519 from nealrichardson/reticulate-followup and squashes the following commits:

4336e2ad0 <Neal Richardson> Fix on.exit calls

Authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>