ARROW-8216: [C++][Compute] Filter out nulls by default
In C++ and R, Filter's default is now to drop from output when a selection slot is null. In Python, `Array.filter` now has a keyword argument `drop_nulls=True`. In R, since the expected behavior is the opposite, the argument is `keep_na = TRUE`. Old behavior is preserved in c_glib; I'm not familiar enough with that binding to add an optional parameter in an idiomatic fashion (@mrkn ?)

Closes #6732 from bkietz/8216-Filtering-returns-all-mis

Lead-authored-by: Benjamin Kietzman <bengilgit@gmail.com>
Co-authored-by: Kenta Murata <mrkn@mrkn.jp>
Co-authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>