ARROW-8305: [Java] ExtensionTypeVector should make sure underlyingVector not null
Related to [ARROW-8305](https://issues.apache.org/jira/browse/ARROW-8305).

In ExtensionTypeVector seems not check nullability for underlyingVector, and it will throw exception when it is null in some api like accept/reset/getField etc.

it's better to do null check when create ExtensionTypeVector rather than when really use it.

Closes #6798 from tianchen92/ARROW-8305

Authored-by: tianchen <niki.lj@alibaba-inc.com>
Signed-off-by: David Li <li.davidm96@gmail.com>