ARROW-8237: [Python][Documentation] Review Python developer documentation, add Dockerfile showing minimal source build with conda and pip/virtualenv
This provides Dockerfile setups to build both with Fedora 31 (very recent) and Ubuntu 18.04, both with conda and pip/virtualenv methods

I made the Python developer instructions a bit more minimal. It's unclear that the average developer needs to build Gandiva and Flight, for example. It's fairly clear how to enable these optional components if they are desired.

Closes #6842 from wesm/ARROW-8237

Authored-by: Wes McKinney <wesm+git@apache.org>
Signed-off-by: Wes McKinney <wesm+git@apache.org>