ARROW-8708: [CI] Utilize github actions cache for docker-compose volumes
Github actions has increased the repository level cache size to 5GB which can be enough for the C++ based builds. With a higher compression level the ubuntu cache consumed less than 300MB of disk.

It also worth mentioning, that the github actions caches are evicted once a day even if the 5GB limit has been reached. So basically we don't need to worry about the cache size at all.

I set up each job to use a platform specific cache key ~prefixed with docker and postfixed with the actual git reference. This means that `docker-ubuntu-18.04-refs/pull/<pr>/merge` key is used in case of a pull request and `docker-ubuntu-18.04-master` on master. Also set a restore key to the master key as a fallback to have a cache hit on new pull requests.~

Caching has builtin support for [branch hierarchy](https://help.github.com/en/actions/configuring-and-managing-workflows/caching-dependencies-to-speed-up-workflows#example-of-search-priority).

Closes #7081 from kszucs/gha-ccache

Authored-by: Krisztián Szűcs <szucs.krisztian@gmail.com>
Signed-off-by: Krisztián Szűcs <szucs.krisztian@gmail.com>