ARROW-8481: [Java] Provide an allocation manager based on Unsafe API
This is in response to the discussion in https://github.com/apache/arrow/pull/6323#issuecomment-614195070

In this issue, we provide an allocation manager that is capable of allocation large (> 2GB) buffers. In addition, it does not depend on the netty library, which is aligning with the general trend of removing netty dependencies. In the future, we are going to make it the default allocation manager.

Closes #6956 from liyafan82/fly_0416_unsf

Lead-authored-by: liyafan82 <fan_li_ya@foxmail.com>
Co-authored-by: emkornfield <emkornfield@gmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>