ARROW-3827: [Rust] Implement UnionArray Updated
Implements `UnionArray`.  This PR was getting too big as it was so I will address the following as follow up PR's:

[ARROW-8546](https://issues.apache.org/jira/browse/ARROW-8546)
[ARROW-8547](https://issues.apache.org/jira/browse/ARROW-8547)
[ARROW-8576](https://issues.apache.org/jira/browse/ARROW-8576)

Note that I believe that we should split up `array.rs` and use separate underlying modules for each of the array types, keeping the `Array` trait and some common functionality in `array.rs`.

Having a single file for all the array types worked fine to begin with but now it's getting bloated.  This PR assumes we agree on this (at which point I'll open a follow up PR to move the other array types), if you disagree I can update this to move `UnionArray` back into `array.rs`, etc.

Closes #7004 from paddyhoran/union_2

Authored-by: Paddy Horan <paddyhoran@hotmail.com>
Signed-off-by: Neville Dipale <nevilledips@gmail.com>