ARROW-8822: [Rust] [DataFusion] Add InMemoryScan to LogicalPlan
Add InMemoryScan to LogicalPlan. I forgot to add this one when adding CsvScan and ParquetScan and this one turned out to be a larger change because the memory data source was a bit neglected.

This PR introduces a new Physical Plan for reading partitioned in-memory data sources and updates the MemTable to delegate to it, adding support for partitioning as a side-effect.

Closes #7203 from andygrove/ARROW-8822

Authored-by: Andy Grove <andygrove73@gmail.com>
Signed-off-by: Andy Grove <andygrove73@gmail.com>