ARROW-9343: [C++][Gandiva] CastInt/Float from string functions should handle leading/trailing white spaces
Also refactored the code to remove the parse_float helper function from context. This was done earlier to save the constructor cost when the stringconverter object from arrow/util/parsing was needed to be constructed first. Now that it has changed to parametric functions, its not needed to keep them in the executioncontext.

Closes #7653 from projjal/parsefloat and squashes the following commits:

073061d3a <Projjal Chanda> added the string to error message
2b089dd95 <Projjal Chanda> trim strings before parsing float/int
14de7e61d <Projjal Chanda> refactored gandiva parse float code

Authored-by: Projjal Chanda <iam@pchanda.com>
Signed-off-by: Praveen <praveen@dremio.com>