ARROW-9297: [C++][Parquet] Support chunked row groups in RowGroupRecordBatchReader
In RowGroupRecordBatchReader, each RowGroup is first read as a Table then RecordBatches are yielded from those tables. Previously the Table's lifetime was bounded by the `ReadNext(shared_ptr<RecordBatch>*)` call, which required that a single RecordBatch be yielded per row group. An error was raised in other cases, including the case where a column of string needed chunking to avoid overflowing the 2GB maximum buffer capacity of a single chunk.

- RowGroup Tables are now kept alive across calls to ReadNext until they are fully drained of RecordBatches.
- Repeated Table reading code has been replaced by reuse of RowGroupReader.
- RowGroupRecordBatchReader is now a thin wrapper around an `Iterator` which handles flattening of a sequence of RowGroups (which are sequences of RecordBatches) into a single sequence of RecordBatches

Closes #7704 from bkietz/9297-parquet-recordbatchreader-should-chunk

Lead-authored-by: Benjamin Kietzman <bengilgit@gmail.com>
Co-authored-by: Wes McKinney <wesm@apache.org>
Signed-off-by: Wes McKinney <wesm@apache.org>