ARROW-9501: Add logic in timestampdiff() when end date is last day of…
… a month

timestampdiff(month, startDate, endDate) returns wrong result in Gandiva when the endDate < startDate and endDate is the last day of the month. An additional month is said to have passed when the end day is greater than or equal to the start day, but this does not hold true for dates which are last days of the month.

Case in point, if startDate = 2020-01-31, endDate = 2020-02-29, previously timestampdiff() returned 0, but the correct result should be 1.

Closes #7782 from sagnikc-dremio/diff and squashes the following commits:

665fa27a7 <Sagnik Chakraborty> ARROW-9501: Add logic in timestampdiff() when end date is last day of a month

Authored-by: Sagnik Chakraborty <sagnikc@dremio.com>
Signed-off-by: Praveen <praveen@dremio.com>