ARROW-10163: [Rust] [DataFusion] Add DictionaryArray coercion support
This PR  adds:
1. Basic `DictionaryArray` coercion (not cast) support in DataFusion
2. A test in `sql.rs` demonstrating basic operations using DataFusion on `DictionaryArray` arrays

Note that the performance operating on `DictionaryArrays` is likely to leave a lot to be desired -- specifically almost any operation will cause the `DictionaryArray` to get unpacked to a normal array reducing most/all of any performance gains.

I plan to add additional performance improvements over time -- but I felt getting queries to run was the first important step.

Closes #8463 from alamb/alamb/ARROW-10159-dictionary-array-coercion-take-3

Authored-by: alamb <andrew@nerdnetworks.org>
Signed-off-by: Neville Dipale <nevilledips@gmail.com>