ARROW-10476: [Rust] Allow string arrays to be built from Option<&str> or Option<String>
Currently, our code-base supports creating a `StringArray` from an iterator of `String`.

However, from arrow's perspective, we should not care if it is a `String` or a `&str`, as long as it can be represented by an `AsRef<str>`. A user sometimes is able to create an iterator of `&str` instead of `String`, and should not have to convert one to the other before passing it to Arrow.

This PR makes this change.

Closes #8575 from jorgecarleitao/array_from_str

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>
Signed-off-by: Neville Dipale <nevilledips@gmail.com>