ARROW-10673: [Rust] [DataFusion] Made sort not collect on `execute`.
Currently, we `collect` and `sort` the record batches from the incoming part of `Sort` on `execute`. However, there is no need to do so: we can postpone that to the stream, like we do for our aggregates. This allow executors to postpone the heavy operation of collecting and sorting to only when they want the first (and only) batch from the stream.

This PR does exactly this: it postpones the execution to when needed.

This code's design is based on HashAggregate, that also uses a one shot channel.

Closes #8730 from jorgecarleitao/sort_post

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>
Signed-off-by: Andrew Lamb <andrew@nerdnetworks.org>