ARROW-10690: [Java] Fix ComplexCopier bug for list vector
When copying a list vector using ComplexCopier, if the target list vector is non-empty, the result vector is incorrect. For example, if src = [[1,2], [3, 4]], tgt = [[5, 6], [7, 8]], copying src to tgt gives tgt = [[5, 6, 1, 2], [3, 4]] instead of [[1,2], [3, 4]]
Similary when using copyFrom(/Safe) method if the vector at the index is non-empty, the listElement is appended instead. For above src and tgt vector, tgt.copyFrom(0, 0, src) gives [[5, 6, 1, 2], []] instead of [[1,2], [7, 8]]

Closes #8742 from projjal/listcomplexcopier and squashes the following commits:

15313199f <Projjal Chanda> move condition from list writer writer to list vector
71112eb48 <Projjal Chanda> fix complex copier

Authored-by: Projjal Chanda <iam@pchanda.com>
Signed-off-by: Praveen <praveen@dremio.com>