ARROW-9728: [Rust] [Parquet] Nested definition & repetition for structs
This is the change to enable writing (and reading) nested structs correctly, ala `<struct<struct<primitive>>>`.

The change introduces a busy `LevelInfo` struct, which is mostly useful for tracking changes across nested lists.
I opted to get structs right first, then refocus on lists, so our list IO support is still broken for now.

I also picked up an issue with our dictionary write support, in that some of the indexing didn't look right. i'm going to work on that in a separate commit, as I squashed all my work on this PR into one commit.

There's a nested struct test that's failing in this PR, but that's because I need #8739 in order for that test to pass.

I haven't focused on making this performant as yet, so there might be plenty of room to improve performance. I mainly focused on getting the nesting arithmetic correct. I think we should be able to paralleli(s|z)e the logic per array, but alas.

In future PR's, I'll work on list support.

Closes #8792 from nevi-me/ARROW-9728

Authored-by: Neville Dipale <nevilledips@gmail.com>
Signed-off-by: Andrew Lamb <andrew@nerdnetworks.org>