ARROW-11114: [Java] Fix Schema and Field metadata JSON serialization
Based on [this commit](https://github.com/apache/arrow/commit/62f322c230cd71ba7256c4fccdcfdeaefa550ad6), it seems that the correct format for metadata serialization is `List<Map<String, String>>` since it was decided to deserialize it in this format.

There are several ways to accomplish this, but this PR takes a less-impactful approach of defining the correct getter for Jackson to use. Other approaches could be more aggressive and propagate this change throughout the codebase, modifying the `Schema#getCustomMetadata` and `Field#getMetadata` method to return `List<Map<String, String>>`, but at that point it would probably make more sense to get rid of the notion of "converting" metadata entirely.

Closes #9088 from nbruno/ARROW-11114

Authored-by: Nick Bruno <nbruno@users.noreply.github.com>
Signed-off-by: liyafan82 <fan_li_ya@foxmail.com>