ARROW-1195: [Rust] [DataFusion] Add accessor methods for Parquet and CSV TableProviders
For projects that depend on the DataFusion logical plan, there is a breaking change currently compared to 2.0.0 where it is not possible to introspect the TableScan node for CSV and Parquet files to get the file path and other meta-data (such as delimiter for CSV).

I propose adding public methods to the specific providers so that other crates can downcast to specific implementations in order to access this data.

My specific use case is the ability to serialize a DataFusion logical plan.

Closes #9149 from andygrove/expose-table-provider-data

Authored-by: Andy Grove <andygrove73@gmail.com>
Signed-off-by: Andy Grove <andygrove73@gmail.com>