ARROW-11461: [Go][Flight] Some cleanup for flight, Fix Schema bytes
Refactoring around the ipc writer and reader to remove some duplication and improve the interfaces.

Also fixing the interactions with `GetSchema` calls and `GetFlightInfo` schema calls to handle the way that existing servers and clients actually implement this.

Currently pkg.go.dev says no one is importing the flight package, so I feel comfortable with this as a breaking change for now, personally. Unless others would prefer to have this not be a breaking change and just deprecate the existing stuff (which would be a bit difficult and cause an import cycle between flight and ipc modules).

Closes #9389 from zeroshade/cleanupflight

Authored-by: Matthew Topol <mtopol@factset.com>
Signed-off-by: Sebastien Binet <binet@cern.ch>