ARROW-11788: [Java] Fix appending empty delta vectors
This PR fixes a bug where appending an empty list vector fails with a `NullPointerException`. Instead of attempting to process the delta vector, we just return early with the unmodified target vector, since there is nothing to append to the target vector from the delta vector. In addition, it fixes cases where appending an empty delta vector would've caused the same NPE, or where it'd be an optimization to not attempt processing an empty delta vector. Unit tests are added for all supported `VectorAppender` vector types.

Closes #9581 from nbruno/ARROW-11788

Authored-by: Nick Bruno <nbruno@users.noreply.github.com>
Signed-off-by: liyafan82 <fan_li_ya@foxmail.com>