ARROW-11899: [Java] Refactor the compression codec implementation into core/Arrow specific parts
This issue is in response to the discussion in https://github.com/apache/arrow/pull/8949/files#r588049088

We want to refactor the compression codec related code into two parts: one for the core compression logic, and the other for Arrow specific logic.

This will make it easier to support other compression types.

Closes #9769 from liyafan82/fly_0322_ref

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>