ARROW-12235: [Rust][DataFusion] LIMIT returns incorrect results when used with several small partitions
Prior to this PR DataFusion could return incorrect results for queries that use LIMIT and have multiple partitions

I noticed when I was running some queries locally that `LIMIT` was not behaving correctly. For my case, a query with `LIMIT 10` was always returning zero rows.

It turns out only the first partition to limit was ever consulted. This PR fixes that bug and adds a test

Closes #9924 from alamb/alamb/limit_or_not

Authored-by: Andrew Lamb <andrew@nerdnetworks.org>
Signed-off-by: Andrew Lamb <andrew@nerdnetworks.org>