ARROW-11752: [R] Replace usage of testthat::expect_is()
`testthat::expect_is` is now deprecated - this PR replaces uses of it with alternative functions.  When updating `expect_dplyr_error`, I fixed an issue with its usage which led to one of the tests failing as it no longer gives the expected error, so have set this test to skip.

Closes #9909 from thisisnic/arrow-11752

Lead-authored-by: Nic Crane <thisisnic@gmail.com>
Co-authored-by: Nic <thisisnic@gmail.com>
Signed-off-by: Krisztián Szűcs <szucs.krisztian@gmail.com>