ARROW-11950: [C++][Compute] Add unary negative kernel
This pull request adds the Negate arithmetic compute kernel for integral and floating-point types.  The NegateChecked version is not implemented for unsigned integral types and overflow behavior is consistent with equivalent Add/Subtract operations.  The Negate kernels are registered as "negate" and "negate_checked".

This PR also extends support for unary arithmetic compute kernels and tests.

@bkietz please review

Closes #10113 from edponce/ARROW-11950-Compute-Add-unary-negative-kernel

Authored-by: Eduardo Ponce <edponce00@gmail.com>
Signed-off-by: Benjamin Kietzman <bengilgit@gmail.com>