ARROW-12522: [C++] Add ReadRangeCache::WaitFor
This was split out of ARROW-11883 since it may also be useful to test with ARROW-11772.

This adds a method to get a Future<> from a ReadRangeCache so it can be easily used in an async context. Also, it adds a config flag to make the cache not perform readahead so that readahead can be handled at a different layer of the stack.

Closes #10145 from lidavidm/async-cache

Authored-by: David Li <li.davidm96@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>