ARROW-13044: [Java] Change UnionVector and DenseUnionVector to extend AbstractContainerVector
Currently UnionVector and DenseUnionVector do not extend any base class, only implement the FieldVector interface. This change makes these vectors extend AbstractContainerVector, which is a subclass of BaseValueVector. This allows a the union vectors to be used in extension types as the underlying vector storage.

The current naming of child fields in the union vector ignores the original name and generates a new name based on type (and typeid). This has been changed to respect the original Field name if given, and only if the name is empty it will generate a new name as before. Some current tests add child Fields with empty names, so this preserves the original behavior.

Included tests to verify the added methods from AbstractContainerVector.

Closes #10513 from BryanCutler/union-extend-BaseValueVector-ARROW-13044

Authored-by: Bryan Cutler <cutlerb@gmail.com>
Signed-off-by: Bryan Cutler <cutlerb@gmail.com>