ARROW-11889: [C++] Add parallelism to streaming CSV reader
This converts the parser & decoder into map functions and then creates the streaming CSV reader as an async generator.  Parallel readahead is then added on top of the parser/decoder to allow for parallel reads.

One thing that is lost at the moment is the ability to encounter a parsing error and then continue.  There was a python test that read in the first block, failed to convert the second block, and then successfully read in a third block.  I'm not sure if that restart behavior is important but if it is I can look into adding it.

Another thing that could be investigated in the future is combining the file readers and table readers more.  They already share some components but the parsing and decoding logic, while basically the same, is handled very differently.  The only real difference is that the table reader saves all the parsed blocks for re-parsing and the streaming reader does not.

Closes #10568 from westonpace/feature/ARROW-11889--c-add-parallelism-to-streaming-csv-reader

Authored-by: Weston Pace <weston.pace@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>