ARROW-13982: [C++] Don't stall in async scanner if a fragment generates no batches
Since reordering uses metadata on the batches, we have to emit at least one batch for every fragment, else we'll wait forever for a batch that never arrives. This adds a utility generator to emit an empty batch if the underlying generator is empty (that way we don't need to fix each file format individually).

Closes #11145 from lidavidm/arrow-13982

Authored-by: David Li <li.davidm96@gmail.com>
Signed-off-by: David Li <li.davidm96@gmail.com>