ARROW-13772: [R] Binding for median aggregation
This adds a binding for `median()` in `dplyr::summarise()`. This also adds a binding for `quantile()` but only for a single probability per function call (`length(probs) == 1`). With both bindings, the results are approximate, calculated using the tdigest algorithm. The user is warned once per session that the results are approximate.

Closes #11018 from thisisnic/ARROW-13772_median_hash

Lead-authored-by: Ian Cook <ianmcook@gmail.com>
Co-authored-by: Nic Crane <thisisnic@gmail.com>
Signed-off-by: Ian Cook <ianmcook@gmail.com>