ARROW-13847: [Java] Avoid unnecessary collection copies
There are several occurences in Arrow Java codebase where temporary
collections are being created and immediately discarded like when going
over a sequence of elements, or when making a copy of a collection
provided as an argument to a constructor.

This change adds a couple of methods to Collections2 and AutoCloseables,
along with some changes to the codebase to remove those extra copies.

Closes #11063 from laurentgo/laurentgo/avoid-copies

Lead-authored-by: Laurent Goujon <laurent@apache.org>
Co-authored-by: emkornfield <emkornfield@gmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>