ARROW-13623 [R] write_csv_arrow update to follow the signature of readr::write_csv
@thisisnic @jonkeane & @paleolimbot This is my first attempt at the task. Not convinced this is the best approach for exposing the args to the user (in prep for hooking them up to the C++ functionality).
I plan on adding more unit test for this PR.
I'd be tempted to suggest we start using some sort of `lifecycle` approach. In such a case, we could have a function called `in_progress()` which would indicate we are working on exposing functionality. Similarly the `tidyverse` uses, for example, `lifecycle::deprecated()`

Closes #11668 from dragosmg/ARROW-13623_write_csv_arrow_signature_update

Lead-authored-by: Dragos Moldovan-Grünfeld <dragos.mold@gmail.com>
Co-authored-by: Dragoș Moldovan-Grünfeld <dragos.mold@gmail.com>
Signed-off-by: Jonathan Keane <jkeane@gmail.com>