ARROW-5599: [Go] Migrate array.{Interface,Record,Column,Chunked,Table} to arrow.{Array,Record,Column,Chunked,Table}
Aliases are introduced in the array package so that existing consumers do not get broken by the migration. Deprecation messages are added to the comments so that they are warned that the aliases will be removed in v8.

Also adds an interface `arrow.ArrayData`. I personally like this updated naming scheme and it makes things easier to understand.

Closes #11832 from zeroshade/arrow-5599

Authored-by: Matthew Topol <mtopol@factset.com>
Signed-off-by: Matt Topol <zotthewizard@gmail.com>