ARROW-15376: [Go][Release] cpu_arm64 needs +build comment
The `//go:build` syntax being used in place of `// +build` was introduced in go1.17. Our CI runs with go1.16 so as long as we want to support go1.16 we need both the `//go:build` and `// +build` . Currently only the `cpu_arm64.go` file has the `//go:build` syntax which is only considered when building on an arm64 machine. That's why it only failed on the Apple M1 verification.

Adding the `// +build arm64` line will allow it to pass verification.

Closes #12214 from zeroshade/arrow-15376-release

Authored-by: Matthew Topol <mtopol@factset.com>
Signed-off-by: Krisztián Szűcs <szucs.krisztian@gmail.com>