ARROW-15331: [Go][Parquet] Add pqarrow package for direct Parquet <--> Arrow conversion
This is the last large portion of code to get merged. anything after this is individual bug fixes or improvements. With the exception of features that aren't yet implemented in Go (like Dictionary Arrays and such) this brings the Go Parquet library pretty much to mostly feature parity with the C++ Parquet Library.

Closes #12149 from zeroshade/parquet-pqarrow

Authored-by: Matthew Topol <mtopol@factset.com>
Signed-off-by: Matt Topol <zotthewizard@gmail.com>