ARROW-15631: [Packaging][RPM] Add major version to libs packages
We always use XXX-libs such as arrow-libs for all versions for now. It
has the following problems:

1. "dnf update" is failed after we release a new version
   until downstream packages are rebuilt for the new version.
2. Users can't use different major versions such as 6.0.0 and 7.0.0
   in the same system.

Example for 1.: Groonga https://groonga.org/ uses Apache Arrow and it
provides RPM packages built with arrow-libs. If Groonga's package is
built with arrow-libs for 6.0.0, "dnf update" is failed when we
release arrow-libs for 7.0.0. Because "dnf update" tries to update
arrow-libs to 7.0.0 from 6.0.0 but Groonga's packages depend on
arrow-libs for 6.0.0 not 7.0.0. This isn't resolved until Groonga's
packages are rebuilt with arrow-libs for 7.0.0. If we provide
arrow6-libs for 6.0.0 and arrow7-libs for 7.0.0, "dnf update" isn't
failed without rebuilding Groonga's packages with arrow7-libs.

Closes #12385 from kou/packaging-rpm-versioned-arrow-libs

Authored-by: Sutou Kouhei <kou@clear-code.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>