ARROW-14893: [C++] Allow creating GCS filesystem from URI
This is mostly adapted from s3. One place that it differs
is it doesn't allow for username and password. The only thing
that I could see supporting in the future is taking username as
user to impersonate, but we can see if there is demand for that feature.

I also added in two useful features in s3:
- Option for setting a default location to create a bucket in.
- Default key-value metadata to use for object creation if none
  is specified.

Also fixed a typo in s3fs error message.

Closes #12475 from emkornfield/gcs_python

Lead-authored-by: Micah Kornfield <micahk@google.com>
Co-authored-by: emkornfield <emkornfield@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>