ARROW-15772: [Go][Flight] Server Basic Auth Middleware/Interceptor wrongly base64 decode
The proposed fixed made the server auth works for both actual go implementation and other. I am not sure if the base64 padding should be used but in all cases having the server being able to read both seems better (being liberal with the inputs, strict with the output).

An other fix would be to change the Go implementation of the client to use the padded base64 (in `client.go` l.299). But I think making the server more robust is better.

Closes #12503 from Corentin-pro/auth-base64-fix

Authored-by: Corentin <corentin-pro@mail.com>
Signed-off-by: Matthew Topol <mtopol@factset.com>