ARROW-15817: [R] Use TableSourceNode instead of InMemoryDataset
This allows us to use the query engine without `ARROW_DATASET=ON`, and it's nice to just keep around the object that was provided instead of transforming it, which makes printing query objects a little more intuitive. There may be other optimizations or cleanups we could do as well.

Other changes: added head/tail methods for RecordBatchReader and fixed an edge case (`head(x, 0)` works but was previously rejected as invalid input).

Closes #12533 from nealrichardson/table-query

Authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>