ARROW-15947: [R] rename_with s3 method for arrow_dplyr_query
Created a simple version of `rename_with` that applies the function to the current names of the .data argument and passes the result to `rename`.

This was a quick writeup on my end and passed a few smoke tests. I believe all the functions I've used in the new code are from base R or current imports like rlang, and the magic in editing the arrow data query is handled by `rename`.

Let me know what next steps can be, like writing test cases, etc.

Closes #12642 from markromanmiller/dplyr_rename_with

Lead-authored-by: MrMallIronmaker <725mrm@gmail.com>
Co-authored-by: Jonathan Keane <jkeane@gmail.com>
Signed-off-by: Jonathan Keane <jkeane@gmail.com>