ARROW-16193: [Go] Replace CPU discovery package with golang.org/x/sys/cpu module
The package deleted in this commit is an outdated copy of a package that
used to be an internal package in the Go standard library, but since
then it has been moved to the publicly accessible golang.org/x/sys
module. Since that module is already used, this commit does not even add
a new dependency.

Other than removing unnecessary code it also makes CPU feature discovery
more consistent with other projects in the Go ecosystem. For example it
is no longer necessary to pass an environment variable to enable usage
of ARM64 NEON SIMD instructions.

Closes #12764 from brancz/go-cpu-discovery

Authored-by: Frederic Branczyk <fbranczyk@gmail.com>
Signed-off-by: Matthew Topol <mtopol@factset.com>