MINOR: [CI] update to nightly reports individual job URLs
We currently have around 86K branches on the crossbow repository. This causes some direct problems on browsing and the email reports queries sometimes failing to open.

There are some initiatives to reduce the number of branches (https://github.com/ursacomputing/crossbow/pull/12) both to improve the browsing experience and to avoid possible issues with Github.

One small improvement I propose is to update the URLs on our nightly report emails for individual tasks to avoid a search on github that might timeout.
Instead of using:

https://github.com/ursacomputing/crossbow/branches/all?query=nightly-tests-2022-04-14-0-github-test-build-vcpkg-win

Using the direct branch URL to avoid the search:

https://github.com/ursacomputing/crossbow/tree/nightly-tests-2022-04-14-0-github-test-build-vcpkg-win

Closes #12882 from raulcd/change-report-job-urls

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>