ARROW-15705: [JavaScript] Allowing appending null on children in a StructBuilder
When trying to add a null value to a StructBuilder, the code previously only modified the null-bitmap.
I believe this breaks the spec as it results in the child arrays being a different length to the Struct array (and the child arrays should have the null value appended to their values).

Closes #12451 from Alfred-Mountfield/js-struct-initialisation

Lead-authored-by: Alfred Mountfield <alfie.mountfield@gmail.com>
Co-authored-by: Dominik Moritz <domoritz@gmail.com>
Signed-off-by: Dominik Moritz <domoritz@gmail.com>