ARROW-16136: [Gandiva][C++] Fix problem of the huge size of AddMappings function
The method that add the stubs functions in Gandiva to be exported is huge and it is leading to some errors in the repository’s lint builds in CI, as it showed here: [ARROW-15661: [Gandiva][C++] Add SHA512 function · apache/arrow@dbc615a](https://github.com/apache/arrow/runs/5532621946?check_suite_focus=true)

It is necessary to break that method in smaller components or disable the lint verification for this specific file.

That error blocks the addition of new functions that must be created in stubs files.

Closes #12804 from Johnnathanalmeida/fix/refactoring-AddMappings-function

Authored-by: Johnnathan <johnnathanalmeida@gmail.com>
Signed-off-by: Pindikura Ravindra <ravindra@dremio.com>