ARROW-16456: [Go] Fix RecordBuilder UnmarshalJSON when extra fields are present
If RecordBuilder.UnmarshalJSON encountered an unknown field in the JSON document; it wouldn't consume the next JSON token representing the value.

Fix this by manually calling `dec.Token()` and allowing the next call to `dec.Token()` to return the following key.

Note that this only handles the case where the ignored value is a simple type, and not a nested array or object. I'm not sure what the correct fix for that would be.

Closes #13065 from phillipleblanc/phillip/unmarshal-json-bug

Authored-by: Phillip LeBlanc <phillip@leblanc.tech>
Signed-off-by: Matthew Topol <mtopol@factset.com>