ARROW-16592: [C++][Python][FlightRPC] Finish after failed writes (#13191)
Currently a failed write (due to the server sending an error,
disconnecting, etc.) will raise an uninformative error on the
client. Prior to the refactoring done in Arrow 8.0.0, this was
silently swallowed (so clients would not get any indication until
they finished writing). In 8.0.0 instead the error got propagated
but this led to confusing, uninformative errors. Instead, tag this
specific error so that the client implementation knows to finish
the call and get the actual server error.

(gRPC doesn't give us the actual error until we explicitly finish
the call, so we can't get the actual error directly.)

Authored-by: David Li <li.davidm96@gmail.com>
Signed-off-by: Yibo Cai <yibo.cai@arm.com>