ARROW-16948: [C++] Benchmark Aggregates Fails To Compile After Aggregate Updates (#13489)
This PR is a FIX for compilation issue in `aggregate_benchmark.cc` and possibly didn't covered by my previous PR on Aggregate simplification (https://issues.apache.org/jira/browse/ARROW-16549). 

PS: This wasn't compiled on Mac M1. In addition also think there could be a CI improvement that can be later on added to cover this. Open for discussion.

Authored-by: Vibhatha Abeykoon <vibhatha@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>