ARROW-16989: [C++] Substrait ProjectRel is interpreted incorrectly (#13528)
A Substrait ProjectRel *appends* columns to the dataset, while Acero's project node replaces them (emit clauses are instead used to remove or swizzle columns). This PR prefixes the current columns in the project node to make the two compatible.

I don't think a declaration includes information about the number of columns it generates, so I had to refactor a little bit to have relation ToProto return a struct of the declaration and the number of columns, in order to know how many columns to replicate. I expect this struct to grow in importance and features when ARROW-16986 is addressed.

Authored-by: Jeroen van Straten <jeroen.van.straten@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>