MINOR: [R] remove duplication about hive-style file paths
Reading the vignette about the datasets, it seems that the part about having self-describing file paths is repeated.

This PR removes the second time this is mentioned and adds the link to the Hive project when it's first mentioned.

Another small detail is that the months in the dataset (at least in the S3 bucket) use a single digit (e.g., `1` for January) while in the section removed by this PR they are listed with 2 digits (`01` for January).

Closes #13844 from fmichonneau/rm-hive-duplication

Authored-by: François Michonneau <francois.michonneau@gmail.com>
Signed-off-by: Nic Crane <thisisnic@gmail.com>