ARROW-18088: [CI][Python] Fix pandas master/nightly build failure related to timedelta (#14460)
Pandas 2.0+ will now preserve the unit of a timedelta64/datetime64 numpy array that is passed to a DataFrame constructor, instead of always coercing to nanoseconds. But pyarrow will still use nanoseconds in the conversion to pandas, giving data type mismatches in the tests. For now, ensuring we always create pandas.DataFrames with datetime64[ns] across pandas versions. Longer term solution is ARROW-18124

Authored-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>
Signed-off-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>