ARROW-17509: [C++] Simplify async scheduler by removing the need to call End (#14524)
This call to end was a foot-gun because failing to call End on a scheduler would lead to deadlock.  This led to numerous situations where one would accidentally fail to call End because of some kind of error or exceptional behavior.

Now the call to end is no longer required.  The scheduler has also been simplified by breaking into three different concepts.  The root scheduler runs a bunch of tasks until there are no more tasks and then automatically ends.

A throttle can wrap a scheduler and throttle task execution, deferring tasks to a queue if there is no space for new tasks.

A task group tracks a group of tasks and runs a finish callback when all the tasks in that group have finished.

Lead-authored-by: Weston Pace <weston.pace@gmail.com>
Co-authored-by: David Li <li.davidm96@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>