ARROW-17288: [C++] Adapt the CSV file format to the new scan API (#14663)
This also adds a utility method arrow::compute::DeclarationToReader.

This does not fully satisfy ARROW-17288 as it only adapts the CSV file format.  However, this was the most complicated case, and I had to convert some of the common test utilities as well.  So I am hopeful the other formats will be more straightforward.

Authored-by: Weston Pace <weston.pace@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>