ARROW-15592: [C++] Add support for custom output field names in a substrait::PlanRel (#14292)
Substrait Plans provide output names in the root relation. This is not completely enabled in the existing Acero-Substrait API and a test case was not there to validate the functionality. To elaborate it's action, a `TableSinkNodeConsumer` is the ideal option. Using this consumer, the PR test shows how the included feature satisfies the requirement. 

PR includes a test case and some minor changes to the existing factories to enable this feature.

Authored-by: Vibhatha Lakmal Abeykoon <vibhatha@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>