MINOR: [Docs][Python] Add example of not nullable type in Schema (#14926)
It took me a while to figure out from the docs how to use a not nullable type in a schema. The inclusion of strings "nullable", "not null" right on the Schema page makes it more easy to find.
Example like this also makes it more clear the tuple Schema takes is shorthand version of Field.

Authored-by: Martin Mokry <martin-kokos@users.noreply.github.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>