ARROW-18353: [C++][FlightRPC] Prevent concurrent Finish in UCX (#15034)
The reader and writer can be used concurrently with each other,
and both may try to close the underlying stream when an error
happens. This can cause one side to stomp on the other's state
inadvertently. When that happens, we may or may not get a crash
or trigger an assertion - but sometimes things just happen to work
out, and the client exits without proper cleanup. The server,
however, doesn't realize this, and so it gets stuck waiting for
the client, preventing the test from finishing.

The longer-term fix would ideally be to let the server forcefully
terminate instead of waiting for clients. (Incidentally, this
problem also affects the gRPC transport, though gRPC is better
at detecting when the client has disappeared.) Here, we add more
assertions and also properly use a lock while doing cleanup.

Requires #15031.

Authored-by: David Li <li.davidm96@gmail.com>
Signed-off-by: David Li <li.davidm96@gmail.com>