GH-33005: [Dev] Update the pull request merge script to work with master or main (#14533)
# Overview
The goal of this pull request is to update the GitHub merge script to work with a repository default branch named `master` or `main`, as part of the effort to rename the Apache Arrow repository's default branch to `main`. The parent Jira ticket can be found [here](https://issues.apache.org/jira/browse/ARROW-15689). 

# Implementation
- In `.github/workflows/dev.yml`, get and pass the default branch name to the merge test script. 
- Added a function named `git_default_branch_name` to `dev/merge_arrow_pr.py`.
- Used `git_default_branch_name` to replace hard-coded usages of `master` in the file.

# Testing
- Ran the `dev/merge_arrow_pr.py` in debug mode to confirm that there are no errors when attempting to merge an existing pull request.
- Ran `test_merge_arrow_pr.py` locally to confirm there are no failures.

# Future Directions
Updated ARROW-18011 to include updating the hard-coded `master` value in `test_merge_arrow_pr.py`.

# Notes
Thank you @ kevingurney for your help with this pull request!
* Closes: #33005

Lead-authored-by: Fiona La <fionala7@gmail.com>
Co-authored-by: lafiona <fionala7@gmail.com>
Co-authored-by: Sreehari Hegden <sreehari.hegden@gmail.com>
Co-authored-by: Antoine Pitrou <pitrou@free.fr>
Co-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Co-authored-by: Kevin Gurney <kgurney@mathworks.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>