GH-15284: [C++] Use DeclarationToExecBatches in Acero plan tests (#15288)
This PR includes a change to the existing test cases in `plan_test.cc` for Acero-based plan evaluation. 

At the moment, using the `StartAndCollect` method and manually passing the `FutureMatcher` and expected results are done and the validation is done on the results obtained from the `StartAndCollect` method. Most of the updated test cases in this PR uses the a common format. 

Replacing that format and using `DeclarationToExecBatches` makes the code more readable and easy to debug. 
### TODO
- [x] Convert other tests with sort options and custom sinks
- [x] Self review 1
- [x] Self review 2
- [x] Check CI Status
- [x] Mark Ready to Review

* Closes: #15284

Lead-authored-by: vibhatha <vibhatha@gmail.com>
Co-authored-by: Vibhatha Lakmal Abeykoon <vibhatha@users.noreply.github.com>
Co-authored-by: Weston Pace <weston.pace@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>