GH-33872: [C++] Remove hacky shared_ptr construction in AppendScalar (#33866)
`AppendScalarImpl` was created for appending scalars based on a contiguous array of `std::shared_ptr<Scalar>`, so when a single `Scalar` reference is passed, a dummy `shared_ptr` is created before the instantiation of `AppendScalarImpl`.

```cpp
  std::shared_ptr<Scalar> shared{const_cast<Scalar*>(&scalar), [](Scalar*) {}};
  return AppendScalarImpl{&shared, &shared + 1, n_repeats, this}.Convert();
```

This seems non-idiomatic [1] and likely inefficient as a control block needs to be heap-allocated for the `shared_ptr`. An alternative way (that simplifies `AppendScalarImpl` as well) is a custom iterator class the wraps the source iterator with an extra dereference operation.

[1] Pointers to `shared_ptr` as class members set a bad-style precedent

### Component(s)

C++

* Closes: #33872

Authored-by: Felipe Oliveira Carvalho <felipekde@gmail.com>
Signed-off-by: David Li <li.davidm96@gmail.com>