GH-32954: [Java][FlightRPC] Remove FlightTestUtil#getStartedServer and bind to port 0 directly (#34357)
### Rationale for this change

Today, the `FlightServer` is created in `FlightTestUtil#getStartedServer`. All tests that need a server go through this function, but this is unnecessary since the server could directly be started on port 0 (the OS will assign a random port later).

### What changes are included in this PR?

All invocations of `FlightTestUtil#getStartedServer` have been removed, and the burden of generating a `Location` object for the server has been placed on the caller.

### Are these changes tested?

Existing tests impacted were run locally to make sure that there no regressions.

### Are there any user-facing changes?

There are no user facing changes. There are no breaking changes to the public API as well.

@ lidavidm
* Closes: #32954

Authored-by: Ramasai <ramasai.tadepalli+3108@gmail.com>
Signed-off-by: David Li <li.davidm96@gmail.com>