GH-34136: [C++] Add a concept of ordering to ExecPlan (#34137)
In addition, it is now possible to bypass the I/O executor on the `record_batch_source`, `exec_batch_source`, and `array_vector_source`.

It is now possible to create a source node from a `gen::Gen` generator.

BREAKING CHANGE: The default executor for `record_batch_source`, `exec_batch_source`, and `array_vector_source` was (erroneously) the plan's CPU executor.  It now defaults properly to the I/O executor.
* Closes: #34136

Authored-by: Weston Pace <weston.pace@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>