GH-34388: [C++] Build core compute kernels unconditionally (#34295)
This includes the core compute machinery in libarrow by default - in addition to all cast kernels and several other kernels that are either dependencies of `cast` (`take`) or utilized in libarrow/libparquet (`unique`, `filter`). The remaining kernels won't be built/registered unless `ARROW_COMPUTE=ON` (note that this would slightly change the option's meaning, as currently, nothing in arrow/compute is built unless it's set).

Initially this was more substantial as the original goal was to build the extra kernels as a shared library (suggested in the orginal issue). After some discussion in the issue thread, I opted not to do that - primarily because I can't personally see the utility of a separate lib here, even ignoring the complexity it introduces. However, there may be a good reason that simply hasn't occured to me.
* Closes: #34388

Lead-authored-by: benibus <bpharks@gmx.com>
Co-authored-by: Ben Harkins <60872452+benibus@users.noreply.github.com>
Co-authored-by: Weston Pace <weston.pace@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>