GH-34422: [R] Expose GcsFileSystem$options (#34477)
Closes https://github.com/apache/arrow/issues/34422

### Rationale for this change

Exposing options on GcsFilesystem objects is useful for testing but may also be useful to R arrow users who are using GCS (gs_bucket). The implementation here is one-way which might be a good point for in discussion in code review.

### Are these changes tested?

I've added two tests which pass.

### Are there any user-facing changes?

Yes, though I wasn't sure if it made sense to document it. With this change, a user can run the following to get the options:

```r
> bucket <- gs_bucket("voltrondata-labs-datasets")
> bucket$base_fs$options
$anonymous
[1] FALSE

$scheme
[1] "https"

$retry_limit_seconds
[1] 15
```
* Closes: #34422

Lead-authored-by: Bryce Mecum <petridish@gmail.com>
Co-authored-by: Nic Crane <thisisnic@gmail.com>
Signed-off-by: Dewey Dunnington <dewey@fishandwhistle.net>