GH-34421: [R] Let GcsFileSystem take a path for json_credentials (#34524)
### Rationale for this change

Existing documentation for this argument was misleading.

### What changes are included in this PR?

A change in functionality, matching tests, and updated documentation are included. `json_credentials` can now either be a literal string containing credentials or a string containing a path to credentials. In the latter case, credentials will be automatically read in from the fileystem.

### Are these changes tested?

Yes

### Are there any user-facing changes?

Yes, though not breaking. This affects user-facing APIs and documentation and is both a bug fix and new functionality.

Closes https://github.com/apache/arrow/issues/34421
Closes https://github.com/apache/arrow/issues/33106
* Closes: #34421

Authored-by: Bryce Mecum <petridish@gmail.com>
Signed-off-by: Dewey Dunnington <dewey@fishandwhistle.net>