GH-35389: [Python] Fix coalesce_keys=False option in join operation (#35505)
### Rationale for this change

During the refactor of the join cython code to use the new `pyarrow.acero` code, I accidentally ignored the `coalesce_keys=False` option. 
This PR restores the previous behaviour (by not passing a custom subset of column names to the HashJoinNode, but relying on its default behaviour to include all fields from left and right data (depending on the join type)).

### Are these changes tested?

Expanded the existing tests to now properly cover the coalesce_keys=False option for all join types.

### Are there any user-facing changes?

Fixes a regression in 12.0, restoring behaviour of 11.0
* Closes: #35389

Authored-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>
Signed-off-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>