GH-35617: [Docs] Current n_buffers use in C API example (#35626)
### Rationale for this change

As described in #35617 the example in docs suffers from a likely copy-and-paste error as the `n_buffers` value references the _parent_ rather that the two children where this is used. `g++-12` spots this and bemoans that assigning to, respectively, `buffers[1]` and then `buffers[1]` and `buffers[2]` is out of bounds.

### What changes are included in this PR?

The example is corrected.

### Are these changes tested?

Yes, locally.  The warnings under `g++-12` go away.

### Are there any user-facing changes?

Not in behavior but the documentation is improved.

* Closes: #35617

Authored-by: Dirk Eddelbuettel <edd@debian.org>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>