GH-14790: [Dev] Avoid extra comment with Closes issue id on PRs (#35811)
### Rationale for this change

We are duplicating the Closes issue_id comment on PRs adding it to both the PR body and a comment. There was some discussion to remove it from the comment.

### What changes are included in this PR?

Remove adding extra comment to PR and check whether `Closes XXX` was already added on the body instead of checking comment.

### Are these changes tested?

Yes, I have tested on my fork, see this PR: https://github.com/raulcd/arrow/pull/79

### Are there any user-facing changes?

No
* Closes: #14790

Lead-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>