GH-35838: [C++] Fix asof join backpresure (#35878)
### Rationale for this change

To fix a bug in asof join backpresure handling.

### What changes are included in this PR?

This is reverting a bug introduced in GH-34391 on this line that breaks asof join backpresure
https://github.com/apache/arrow/pull/34392/files#diff-5493b6ae7ea2a4d5cfb581034c076e9c4be7608382168de6d1301ef67b6c01eeR1410

### Are these changes tested?
No. However code change simply reverts to the state before the bug was introduced in GH-34391 and therefore should be pretty safe (we have tested that the code before GH-34391 is working). In the meantime @ rtpsw is working on adding tests around this but I would like to merge this to unblock internal  issues around this.

### Are there any user-facing changes?

* Closes: #35838

Authored-by: Li Jin <ice.xelloss@gmail.com>
Signed-off-by: Li Jin <ice.xelloss@gmail.com>