GH-36151: [Java] Add `volatile` declaration to `keyPosition` in `ParallelSearcher` (#36152)
### Rationale for this change

Multiple threads can read/write to the `keyPosition` variable. Thus, it should be declared `volatile` so that all threads have a consistent view of the value of this variable.

### What changes are included in this PR?

This PR just marks an instance member `volatile`.

### Are these changes tested?

Existing tests should suffice as no functionality is being changed.

### Are there any user-facing changes?

No, there are no user facing changes. There are no changes to the public API either.
* Closes: #36151

Authored-by: Ramasai <ramasai.tadepalli+3108@gmail.com>
Signed-off-by: David Li <li.davidm96@gmail.com>