GH-36345: [C++] Prefer TypeError over Invalid in IsIn and IndexIn kernels (#36358)
### Rationale for this change

`TypeError` should be returned when `values` and `value_set` have incompatible types in IsIn and IndeIx.
`Invalid` is still returned if the types are compatible but casting the values fails (for example because of overflow or truncation).

### What changes are included in this PR?

When casting between types is not supported, return TypeError instead of Invalid.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

No.

* Closes: #36345

Authored-by: Jin Shang <shangjin1997@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>