GH-35130: [Docs] Document how to become a collaborator to get triage role (#36445)
### Rationale for this change

Since we migrated to GitHub we have been using the collaborator role to allow some contributors to start helping with a triage role.
There has been some discussions on making more clear on the process to become a collaborator.

### What changes are included in this PR?

Add some documentation on how to propose / become a collaborator.

### Are these changes tested?

Will trigger the preview-docs.

### Are there any user-facing changes?

No
* Closes: #35130

Lead-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Co-authored-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>