GH-36619: [Python] Parquet statistics string representation misleading (#36626)
### Rationale for this change

If `Statistics.has_distinct_count` is false, the printout still shows a `distinct_count` value (`0`).

### What changes are included in this PR?

Update the `distinct_count` property  of `Statistics` class to check for `has_distinct_count` value.

### Are these changes tested?

The tests for `distinct_count` are already included in `test_parquet_column_statistics_api` and only needed to be updated.
* Closes: #36619

Authored-by: AlenkaF <frim.alenka@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>