GH-36687: [R] Add correct branch name to autobrew formulae to facilitate local testing (#36689)
### Rationale for this change

It is currently not possible to recreate an autobrew build locally by following the instructions in the comments. This fixes the local copies of the upstream formulas and the instructions so that future debuggers can recreate an autobrew build.

### What changes are included in this PR?

The branch `master` no longer exists and is the default value. This PR adds the revised default branch name ("main").

### Are these changes tested?

No nightly test covers this because this value would be overwritten to test specific commits anyway.

### Are there any user-facing changes?

No.
* Closes: #36687

Authored-by: Dewey Dunnington <dewey@voltrondata.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>