GH-36750: [R] Fix test-r-devdocs on MacOS (#36751)
### Rationale for this change

The test-r-devdocs job is failing. It is failing because we are pinning a version of R that is so old that CRAN no longer serves binaries (#31757), so some of the package builds are failing.

### What changes are included in this PR?

Use the `setup-r-dependencies` action which either installs the correct build dependencies or uses older binary versions to avoid building from source (or both).

### Are these changes tested?

Yes, as part of the test-r-devdocs job.

### Are there any user-facing changes?

No.
* Closes: #36750

Authored-by: Dewey Dunnington <dewey@voltrondata.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>