GH-37307: [Python][CI] Manually skip tests with skip_with_pyarrow_strings marker for nightly dask integration tests (#37324)
### Rationale for this change

Dask added some tests with a custom `@ pytest.mark.skip_with_pyarrow_strings` mark, which ensures to skip some tests when pyarrow is installed. However, that skip doesn't work correctly when running the tests on an installed version of dask with `pytest --pyargs dask.tests`. 
Therefore manually skipping tests with that mark as a workaround.

* Closes: #37307

Authored-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>
Signed-off-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>