GH-37939: [C++] Use signed arithmetic for frame of reference when encoding DELTA_BINARY_PACKED (#37940)
Closes #37939. 

### What changes are included in this PR?

This PR changes values used in the `DELTA_BINARY_PACKED` encoder to signed types. To gracefully handle overflow, arithmetic is still performed in the unsigned domain, but other operations such as computing the min and max deltas are done in the signed domain.

Using signed types ensures the optimal number of bits is used when encoding the deltas, which was not the case before if any negative deltas were encountered (which is obviously common).

### Are these changes tested?
I've included two tests that result in overflow.

### Are there any user-facing changes?
No

* Closes: #37939

Authored-by: seidl <seidl2@llnl.gov>
Signed-off-by: Antoine Pitrou <antoine@python.org>