MINOR: [Docs][Python] Add more text around PyArrow build types (#38522)
### Rationale for this change

It's not very clear that PyArrow will get built in release mode even if the linked Arrow C++ is built in debug mode. I added some text to the docs to make it more clear. An example of a user running into this is https://github.com/apache/arrow/issues/38519.

### Are these changes tested?

No as they're just minor docs changes.

### Are there any user-facing changes?

This adds text to the Python docs.

Authored-by: Bryce Mecum <petridish@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>