GH-39003: [CI][macOS] Don't update Homebrew (#39016)
### Rationale for this change

It's better that we always use the latest Homebrew to check with the latest Homebrew that are used by most users. But it's difficult to maintain.

### What changes are included in this PR?

We don't update Homebrew manually. GitHub hosted GitHub Actions Runners update Homebrew periodically. We depend on it instead of manual `brew update`.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

No.
* Closes: #39003

Authored-by: Sutou Kouhei <kou@clear-code.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>