GH-39656: [Release] Update platform tags for macOS wheels to macosx_10_15 (#39657)
### Rationale for this change

Currently the binary verification for releases fails due to wrong macOS platform version.

### What changes are included in this PR?

Update to the current generated platform tag.

### Are these changes tested?

No, but I've validated this is the corrected generated platform tag for the wheels on the Release Candidate: https://apache.jfrog.io/ui/native/arrow/python-rc/15.0.0-rc1/

### Are there any user-facing changes?

Not because of this change but there was a minimum version of macOS as part of the PR that caused this issue.
* Closes: #39656

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>