GH-40207: [C++] TakeCC: Concatenate only once and delegate to TakeAA instead of TakeCA (#40206)
### Rationale for this change

`take` concatenates chunks when it's applied to a chunked `values` array, but when the `indices` arrays is also `chunked` it concatenates `values` more than once -- one `Concatenate` call with `values.chunks()` for every chunk in `indices`. This PR doesn't remove the concatenation, but ensures it's done only once instead of `indices.size()` times.

### What changes are included in this PR?

 - Adding return type to the `TakeXX` names (-> `TakeXXY`) to makes code easier to understand
 - Adding benchmarks to `TakeCCC` — copied from #13857
 - Remove the concatenation from the loop body (!)

### Are these changes tested?

By existing tests.

### Are there any user-facing changes?

A faster compute kernel.
* GitHub Issue: #40207

Authored-by: Felipe Oliveira Carvalho <felipekde@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>