GH-40672: [Go][Parquet] Add proper build tags for min_max (#40676)
### Rationale for this change
Running something which calls `MinMaxInt32` on a 32-bit architecture was crashing because it wasn't dropping to the noasm solution and was calling assembly designed for a 64-bit architecture. 

### What changes are included in this PR?
Adding the same build constraints to `min_max_noasm.go` as we have on others so that it gets built for 32-bit architectures to fallback to the pure go solution.

* GitHub Issue: #40672

Authored-by: Matt Topol <zotthewizard@gmail.com>
Signed-off-by: Matt Topol <zotthewizard@gmail.com>