GH-41047: [C#] Address performance issue of reading from StringArray (#41048)
### Rationale for this change

The motivation here is to address https://github.com/apache/arrow/issues/41047. There is severe performance drawback in reading a StringArray as value array of a DictionaryArray, because of repeated and unnecessary UTF 8 string decoding.

### What changes are included in this PR?

- Added a new function Materialize() to materialize the values to a list. When materialized, GetString() reads from the vector directly.
- Added test coverage.

### Are these changes tested?

Yes

### Are there any user-facing changes?

No. This change maintains backwards compatibility on the API surface. It is up to the client application to decide whether to materialize the array and gain performance. 

* GitHub Issue: #41047

Authored-by: Keshuang Shen <keshen@microsoft.com>
Signed-off-by: Curt Hagenlocher <curt@hagenlocher.org>