GH-40801: [Docs] Clarify device identifier documentation in the Arrow C Device data interface (#41101)
### Rationale for this change

It is not explicit what the value of the `ArrowDeviceArray::device_id` should be when a given device type has no notion of a device identifier (e.g., there is always only one).

### What changes are included in this PR?

The text was clarified to recommend a value of -1. This was the value already used by Arrow C++.

### Are these changes tested?

No tests needed (documentation)

### Are there any user-facing changes?

No
* GitHub Issue: #40801

Authored-by: Dewey Dunnington <dewey@voltrondata.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>