GH-39645: [Python] Fix read_table for encrypted parquet (#39438)
### Rationale for this change

Currently, if you try to read a decrypted parquet with read_table, passing decryption_properties - in the happy path (pyarrow.data available for import) the reading/decryption of the file fails, as the decryption properties are missing.

### What changes are included in this PR?

Pass through the argument that was intended to have been passed.

### Are these changes tested?

We have tested this locally on an encrypted parquet dataset - please advise on any further testing you would like beyond that and the standard CI.

### Are there any user-facing changes?

Not in any cases where their code was previously working? The intended behaviour for encrypted dataset decryption should start working.

* Closes: #39645

Lead-authored-by: Tom McTiernan <tmct@users.noreply.github.com>
Co-authored-by: Don <tolleybot@gmail.com>
Co-authored-by: Rok Mihevc <rok@mihevc.org>
Signed-off-by: Rok Mihevc <rok@mihevc.org>