GH-40361:  [C++] Make flatbuffers serialization more deterministic (#40392)
### Rationale for this change

This is the start of a PR to address #40361, and in turn #40202, to make metadata in parquet files written by arrow to be identical irrespective of the platform configuration.  This is limited, as platform-specific differences in R or Python versions or compression libraries could still result in differences.

### What changes are included in this PR?

So far I have only made a partial change to part of the metadata serialization.  I need to look at whether other calls to flatbuffers require similar treatment.

### Are these changes tested?

Not yet, this is a draft PR

### Are there any user-facing changes?

No 

* GitHub Issue: #40361

Lead-authored-by: Noam Ross <noam.ross@gmail.com>
Co-authored-by: Bryce Mecum <petridish@gmail.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Co-authored-by: Antoine Pitrou <pitrou@free.fr>
Signed-off-by: Bryce Mecum <petridish@gmail.com>