GH-41945: [Swift] Add interface ArrowArrayHolderBuilder (#41946)
### Rationale for this change

This change adds the implementation of the ArrowArrayHolderBuilder interface which allows appending to Arrays and completing them without needing the generic info.  This is needed for Nested types as well as for the Swift arrow Codable implementation.

### What changes are included in this PR?

Adding the interface and the implementation of the interface.

### Are these changes tested?

Yes, test has been added.

* GitHub Issue: #41945

Authored-by: Alva Bandy <abandy@live.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>