GH-41797: [C++][S3] Remove GetBucketRegion hack for newer AWS SDK versions (#41798)
### Rationale for this change

To get the region a S3 bucket resides on, it is required to issue a HeadBucket request and parse the response headers for a certain header value.

Unfortunately, the AWS SDK doesn't let us access arbitrary headers on successful responses for S3 model requests, which had us implement a workaround by calling lower-level SDK APIs.

However, the SDK recently added a `GetBucketRegion` method on `HeadBucketRequest`, which obsoletes the need for this workaround. We now use this method if the available AWS SDK version is recent enough.

### Are these changes tested?

By existing tests on the various CI configurations.

### Are there any user-facing changes?

No.

* GitHub Issue: #41797

Authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>