GH-42020: [Swift] Add Arrow decoding implementation for Swift Codable (#42023)
### Rationale for this change
This change implements decode for the Arrow Swift Codable implementation.  This allows the data in a RecordBatch to be copied to properties in a struct/class. 

The PR is a bit longer than desired but all three container types are required in order to implement the Decoder protocol.

### What changes are included in this PR?

The ArrowDecoder class is included in this PR along with a class for each container type (keyed, unkeyed, and single).  Most of the logic is encapsulated in the ArrowDecoder with minimal logic in each container class (Most of the methods in the container classes are a single line that calls the ArrowDecoder doDecode methods)

### Are these changes tested?

Yes, a test has been added to test the three types of containers provided by the decoder.

* GitHub Issue: #42020

Authored-by: Alva Bandy <abandy@live.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>