GH-41862: [C++][S3] Fix potential deadlock when closing output stream (#41876)
### Rationale for this change

When the Future returned by `OutputStream::CloseAsync` finishes, it can invoke a user-supplied callback. That callback may well destroy the stream as a side effect. If the stream is a S3 output stream, this might lead to a deadlock involving the mutex in the output stream's `UploadState` structure, since the callback is called with that mutex locked.

### What changes are included in this PR?

Unlock the `UploadState` mutex before marking the Future finished, to avoid deadlocking.

### Are these changes tested?

No. Unfortunately, I wasn't able to write a test that would trigger the original condition. Additional preconditions seem to be required to reproduce the deadlock. For example, it might require a mutex implementation that hangs if destroyed while locked.

### Are there any user-facing changes?

No.

* GitHub Issue: #41862

Authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>