GH-42107: [C++][FS][Azure] Ensure setting BlobSasBuilder::Protocol (#42108)
### Rationale for this change

`BlobSasBuilder::Protocol` is used in `BlobSasBuilder::GenerateSasToken()` and it doesn't have the default value. If we don't specify it explicitly, `GenerateSasToken()` result may be unexpected.

### What changes are included in this PR?

Set `BlobSasBuilder::Protocol` explicitly based on `AzureOptions::blob_storage_scheme`. It's `https` by default but can be `http` when `enable_tls=false` query parameter is specified.

If it's `http`, both of `https` and `http` are accepted as protocol.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

Yes.
* GitHub Issue: #42107

Authored-by: Sutou Kouhei <kou@clear-code.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>