GH-42220: [R] handle vctrs_rcrd extension type in metadata cleaning (#42226)
### Rationale for this change

See https://github.com/apache/arrow/issues/42220. Extension type metadata serialization works slightly differently from the general R metadata path: it uses a vctrs::ptype object, which is 0-length with attributes. The sanitization needs to happen on `attributes()`, not the object being serialized. 

The error was caused because `vctrs_rcrd` has a special `[<-` method, so iterating over it and assigning back in the sanitized version (even if it were the right thing to do) failed.

### What changes are included in this PR?

Sanitize the `attributes()` of vctrs ptypes.

### Are these changes tested?

yes

### Are there any user-facing changes?

other than the bug being fixed, no
* GitHub Issue: #42220

Authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Signed-off-by: Nic Crane <thisisnic@gmail.com>