GH-41952: [R] Turn S3 and ZSTD on by default for macOS (#42210)
Changeup `nixlibs.R` so that we enable S3 and ZSTD by default on CRAN. I've checked this against the CRAN macbuilder to confirm it does work (that's not a guarantee it'll work on other CRAN maintained macOS builders, but a good indication).

It also removes gcs from the list of features we expect to be on and warn folks about if it is not.

### Rationale for this change

So that our builds are more fully featured.

### What changes are included in this PR?

Enable `ARROW_S3` by default when building form source on macOS.

### Are these changes tested?

Existing CI should not fail.

### Are there any user-facing changes?

Getting Arrow from the canonical repository would be more fully featured.
* GitHub Issue: #41952

Lead-authored-by: Jonathan Keane <jkeane@gmail.com>
Co-authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Signed-off-by: Jonathan Keane <jkeane@gmail.com>