GH-43164: [C++] Fix CMake link order for AWS SDK (#43230)
### Rationale for this change

To resolve conflicts with AWS SDK for C++ static variables when linked with libarrow by ensuring correct link order.

### What changes are included in this PR?

- Adjusted `CMakeLists.txt` to set `ARROW_S3_TEST_EXTRA_LINK_LIBS`.
- Ensured `libarrow` is linked before `libaws*` libraries.
- Updated `s3fs_test` configuration to use the new link order.

### Are these changes tested?

I ran the test locally and observed the same result as mentioned. Additionally, I confirmed that if `ARROW_S3` is set to OFF or if the configuration includes `exclude_tests=arrow-s3fs-test`, the test is excluded.

### Are there any user-facing changes?

No.

* GitHub Issue: #43164

Authored-by: Hyunseok Seo <hsseo0501@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>