GH-43202: [C++][Compute] Detect and explicit error for offset overflow in row table (#43226)
### Rationale for this change

There are two places for the offset in a row table to possibly overflow: 1) Encoding columns into a row table; 2) Appending to a row table from another row table.

They are particularly bad because they can cause silent wrong result for the computation.

### What changes are included in this PR?

Detect such overflows in the aforementioned places and report an explicit error when an overflow is detected.

### Are these changes tested?

UT included.

### Are there any user-facing changes?

User code that used to be fake passing could see an explicit error after this change.

**This PR contains a "Critical Fix".**
* GitHub Issue: #43202

Authored-by: Ruoxi Sun <zanmato1984@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>