GH-43284: [Release] Fix version detection timing for bump deb package names on post-12-bump-versions.sh script (#43294)
### Rationale for this change

The script fails to bump deb package names at the moment due to an unmatching if.

### What changes are included in this PR?

Use current_version before bumping the versions in order to match debian packages to be updated.

### Are these changes tested?

It has been tested locally to bump the versions for the debian package names on main for 17.0.0

### Are there any user-facing changes?

No
* GitHub Issue: #43284

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>