GH-43469: [Java] Change the default CompressionCodec.Factory to leverage compression support transparently (#43471)
### Rationale for this change

Add compression support to Flight RPC and others by just including the `arrow-compression` jar in the module path (or classpath).

### What changes are included in this PR?

Change the default compression factory to the new `CompressionCodec.Factory.INSTANCE`, a ServiceLoader-backed singleton that delegates to the best suited available implementation in the module/class path for each codec type.

### Are these changes tested?

yes

### Are there any user-facing changes?

No.
* GitHub Issue: #43469

Authored-by: Costi Ciudatu <ccciudatu@gmail.com>
Signed-off-by: Dane Pitkin <dpitkin@apache.org>